requirements.txt
Flask
newspaper3k
flask-cors